Fix clang-tidy warning

+++ src/core/symbology/qgsfillsymbollayer.cpp
@@ -5589,7 +5589,7 @@ QgsSymbolLayer *QgsRandomMarkerFillSymbolLayer::create( const QVariantMap &prope
     // if we a creating a new random marker fill from scratch, we default to a random seed
     // because seed based fills are just nicer for users vs seeing points jump around with every map refresh
     std::random_device rd;
-    std::mt19937 mt( rd() );
+    std::mt19937 mt( static_cast< int >( rd() ) );
     std::uniform_int_distribution<> uniformDist( 1, 999999999 );
     seed = uniformDist( mt );
   }
